Partially impl SDL3.
May need to review again in the future.

Refactor window init, input, and key mapping logic

Refactored window initialization for GLFW/SDL3 with TryInit method and improved Init logic. Changed time methods to use double for higher precision. Enhanced keyboard key mappings and comments, including NonUS keys. Removed unused input interface methods and moved SDL3-specific event types. Added explicit SDL3 event mapping functions. Updated WindowPlatformFactory for SDL3 support. Improved header includes and added destructors/defaults for interfaces and visitors.

+++ Core/Application/Application.cpp
@@ -37,7 +37,7 @@ namespace CoreEngine {
 		float prevFrameTime = time->GetTimeNow();
 
 		while (mWindow->GetCurrentSeconds() - prevFrameTime < minDeltaTime);
-		float currentTime = mWindow->GetCurrentSeconds();
+		double currentTime = mWindow->GetCurrentSeconds();
 		time->SetDeltaTime(currentTime - prevFrameTime);
 		time->SetTimeNow(currentTime);
